backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
aiohttp==3.11.11
python-multipart==0.0.19
sse-starlette==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
